i20 - Prevent double selection of support cells
Cells now store a list of related cells.
R8 cells store any cells that they specialize in this list.
Any cells specialized by an R8 cell store the R8 cell in this list.

An R8 cell will not specialize a cell if it already has that cell in its
related cells list.